v1 do coinflip

+++ Comandos/BRINCADEIRAS/cntdenum.js
@@ -4,7 +4,7 @@ const db = new QuickDB();
 
 module.exports = {
     name: "contagem", // Coloque o nome do comando
-    description: "『 UTILIDADE 』Inicia a brincadeira de contar.", // Coloque a descrição do comando
+    description: "『 BRINCADEIRAS 』Inicia a brincadeira de contar.", // Coloque a descrição do comando
     type: ApplicationCommandType.ChatInput,
     options: [
         {
@@ -16,9 +16,17 @@ module.exports = {
     ],
 
     run: async (client, interaction) => {
+        let cargos = await db.get(`ArrayCargos.roles`)
+        let valoresGerados = [];
+        for (let index = 0; index < cargos.length; index++) {
+            const element = cargos[index]
+            valoresGerados.push(element)
+        }
+        if (!interaction.member.roles.cache.some(role => valoresGerados.includes(role.id))) return interaction.reply('**VOCÊ NÃO TEM A PERMROLE PARA UTILIZAR ESSE COMANDO!**');
+
         let channelCount = await db.get(`channelCount`);
         let canal = interaction.options.getChannel('canal')
-        if(channelCount === canal.id) return interaction.reply({content: `**❌ | Ei, esse canal já foi configurado como o canal de contagem!**`}).then((msg) => {
+        if (channelCount === canal.id) return interaction.reply({ content: `**❌ | Ei, esse canal já foi configurado como o canal de contagem!**` }).then((msg) => {
             setTimeout(() => {
                 msg.delete();
             }, 10_000);
